pyproject.toml
[project]
name = "meshtastic-bot"
version = "0.1.0"
description = "meshtastic welcome bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "meshtastic>=2.6.4",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["actions"]